+++ torch_geometric/nn/aggr/quantile.py
@@ -115,7 +115,10 @@ class QuantileAggregation(Aggregation):
                 quantile = 0.5 * l_quant + 0.5 * r_quant
 
         # If the number of elements is zero, fill with pre-defined value:
-        mask = (count == 0).repeat_interleave(self.q.numel()).view(shape)
+        repeats = self.q.numel()
+        output_size = repeats * count.numel()
+        mask = (count == 0).repeat_interleave(
+            repeats, output_size=output_size).view(shape)
         out = quantile.masked_fill(mask, self.fill_value)
 
         if self.q.numel() > 1: